Added Grunt, so actual building can go on
Basically, I didn't realize how much I really depend on Grunt to do my
work for me...because I realized that is why developing this web app was
going so slowly. Copy in a Gruntfile, tweak it a bit to match this
project/repo, etc. and suddenly I'm "rolling" again...