Integrated Grunt further, Made schedule directives
I finally got my act together and threw together a little course
schedule widget for myself that I'm hoping to make less cringey very
soon. It makes the courses different colors, and even plots them in the
same color for different days. Currently, this is limited by the number
of colors that I've sort of arbitrarily designated as course-assigning
colors, and the actual placement of the courses could use some work, but
I think that it's going well, so far, at least.

The best part though about the course scheduler is that it makes the
little divs that contain the information with the proper heights and
position for whatever time they began at. Unfortunately, this also means
that the divs are all clustered in the middle, because I have it offset
from 12 am to...12 am, a full day, so there's just no way you'll have a
class in the middle of the night, so I ought to make the range something
like maximum range of the currently chosen courses + 30 minutes of sort
of like buffer space...but that would make things messy, I would think.
Plus, the classes change on a semi-regular fashion so they tend to begin
and end at the same times, so it's difficult to get them to not overlap
such that you can't see the information contained in a div.

I'm considering adding another class for the divs that are longer, one
that makes it such that their z-index is lower than all of the others,
except for when they're being hovered over of course. That way, the
large ones are out of the way in the background, but can still be
accessed.

I also want to add a delete button to the little class items, but I
think that I'll have to add a directive for that. Or a component. Or
both. Whatever.