Modularizing server.js, nearly perfected parser
The server.js file is getting far too large, and to make it more
readable, I'm creating a bunch of module files based on the server.js
file that will replace a lot of the very opaque, impenetrable code with
concise references to module objects. It should be nice.

I have nearly perfected the parser, I think I just need to
a. Come up with a more efficient way of dealing with the proxy server
b. Make the parser a little more concise
c. Make a script that verifies the accuracy of the output to csv
d. Expand the script to include "Notes:", "Waitlist", etc.

I've solved most of the other problems.

I've also added additional directories to the .gitignore file, because I
was using those directories to essentially store junk info. I probably
should have put that in exclude, but w/e.