Started on an index.js file for express app
I decided to break up the node.js folder into several different folders
for different parts of the app, e.g. the Express app, the proxy app,
etc. This should make it a lot cleaner when I finally put it all
together. I'm adding the index.js file to the app/ folder because it'll
automatically run when I require('nodejs/app'), and I don't have to
specify it.