Adding different semesters now.
I really gotta go to dinner, but first I'm going to just say that I've
been incorportating a mechanism that will allow the user to choose a
particular semester, instead of just assuming that they want to schedule
for the fall of 2016, so I've added a select element to index.html for
that.

I also wanted to have different buttons in the index.html file depending
on whether I've selected the production (prod) build or the development
(dev) build, so I have Grunt remove the solely development features of
the site depending on whether I use dev/devs or build/builds.